requirements.txt
Django==4.2.7
Pillow==10.1.0
django-ckeditor==6.7.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
whitenoise==6.6.0